Rename setCalendarData to addCalendarData
